+++ apps/cowswap-frontend/src/common/hooks/useGetExecutedBridgeSummary.ts
@@ -1,3 +1,4 @@
+// TODO: Don't use 'modules' import
 import { useMemo } from 'react'
 
 import { useTokenByAddress } from '@cowprotocol/tokens'