+++ apps/cowswap-frontend/src/entities/bridgeProvider/index.ts
@@ -1,4 +1,5 @@
 export { useBridgeSupportedNetworks, useBridgeSupportedNetwork } from './useBridgeSupportedNetworks'
 export { useBridgeSupportedTokens } from './useBridgeSupportedTokens'
 export { useRoutesAvailability } from './useRoutesAvailability'
+export { useHasHookBridgeProvidersEnabled } from './useHasHookBridgeProvidersEnabled'
 export { BridgeProvidersUpdater } from './BridgeProvidersUpdater'