+++ libs/events/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## [2.1.0](https://github.com/cowprotocol/cowswap/compare/events-v2.0.2...events-v2.1.0) (2025-12-24)
+
+
+### Features
+
+* **bridge:** enable Near bridge provider for smart-contract wallets ([#6618](https://github.com/cowprotocol/cowswap/issues/6618)) ([0da32ab](https://github.com/cowprotocol/cowswap/commit/0da32abcd12183396870055f3cf92bc7b53002a3))
+
+## [2.0.2](https://github.com/cowprotocol/cowswap/compare/events-v2.0.1...events-v2.0.2) (2025-12-16)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/types bumped from ^2.1.0 to ^2.1.1
+
 ## [2.0.1](https://github.com/cowprotocol/cowswap/compare/events-v2.0.0...events-v2.0.1) (2025-11-24)